test(scoped-keys): move api-key principal tests to their own file
CI's golangci-lint typecheck pass continues to report "undefined: testinfra"
on lines inside the new TestAPIKeyPrincipal_* functions while accepting
the same import statement and identical usages in the older tests in
integration_test.go. Local builds and lint pass; the issue is CI-specific
and the symptom matches a typecheck cascade where a real upstream type
resolution failure (likely on urn.PrincipalTypeAPIKey or
APIKeySystemManaged, both new in this PR) gets mis-reported against the
next identifier in the function body.

Splitting the new tests into apikey_principal_test.go gives the CI
typechecker a freshly-named compilation unit to chew on. If the issue
was a stale per-file cache entry, this dodges it. If not, the next
escalation is bumping the GH_CACHE_GO_KEY version constant.
